Modified the tool versions and the test suite
Updated UITest and NUnit to the latest versions.

Modified the Test Suite so that each dialog is part of a separate [Test] instead of the same [Test]. The reason being because if say, Background Location fails but Push Notifications would pass; it wouldn't be able to tell in the current set up because the test would fail before it gets the chance to test Push notifications.